requirements.txt
absl-py==2.1.0
blinker==1.7.0
boto3==1.34.64
botocore==1.34.64
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.10.0
filelock==3.13.1
fire==0.6.0
Flask==3.0.2
fonttools==4.50.0
fsspec==2024.3.0
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.62.1
huggingface-hub==0.21.4
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==5.26.0
psutil==5.9.8
py-cpuinfo==9.0.0
pybboxes==0.1.6
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.24
s3transfer==0.10.1
sahi==0.11.15
scipy==1.12.0
seaborn==0.13.2
shapely==2.0.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.29
urllib3==2.2.1
Werkzeug==3.0.1
yolov5==7.0.13
